algo_api/requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cffi==1.14.3
cryptography==3.2.1
Django==3.1.2
Fabric3==1.14.post1
isort==5.6.4
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
